Created BaseScene class that all scenes inherit from
