+++ PetGame/PetGame/GameScene.swift
@@ -9,12 +9,14 @@
 import SpriteKit
 import GameplayKit
 
-class GameScene: SKScene {
+class GameScene: BaseScene {
     
     private var label : SKLabelNode?
     private var spinnyNode : SKShapeNode?
-    var screenDelegate: ScreenSwitchable?
+    
     override func didMove(to view: SKView) {
+        super.didMove(to: view)
+        
         backgroundColor = SKColor.black
         let background = SKSpriteNode(imageNamed: "background1")
         background.position = CGPoint(x: size.width / 2, y: size.height / 2 )
@@ -36,39 +38,10 @@ class GameScene: SKScene {
         self.view?.addSubview(buttonLobbyScene)
     }
     
-    override func update(_ currentTime: TimeInterval) {
-        super.update(currentTime)
-        
-        self.children.forEach { (child) in
-            if let updatable = child as? Updatable{
-                updatable.update(currentTime)
-            }
-        }
-    }
-    
-    func disableButtons(){
-        guard let subviews = self.view?.subviews else {
-            return
-        }
-        for view in subviews as [UIView] {
-            if let button = view as? UIButton {
-                button.isHidden = true
-                button.isEnabled = false
-            }
-        }
-        
-    }
-    
     @objc func movetoLobbyScene(_ sender:UIButton!)
     {
         //SKView.presentScene(GameScene)
         NotificationCenter.default.post(name: Notification.Name.didRecieveButtonInput, object:nil)
         screenDelegate?.SwitchScreens(string: "LobbyScene")
     }
-
-    
-    
-    
-    
-    
 }
