+++ PetGame/PetGame/WashScene.swift
@@ -10,13 +10,14 @@
 import SpriteKit
 import GameplayKit
 
-class WashScene: SKScene {
+class WashScene: BaseScene {
     
     private var label : SKLabelNode?
     private var spinnyNode : SKShapeNode?
-    var screenDelegate: ScreenSwitchable?
     
     override func didMove(to view: SKView) {
+        super.didMove(to: view)
+        
         backgroundColor = SKColor.black
         let background = SKSpriteNode(imageNamed: "washscenebackground")
         background.position = CGPoint(x: size.width / 2, y: size.height / 2 )
@@ -33,18 +34,6 @@ class WashScene: SKScene {
 
     }
     
-    override func update(_ currentTime: TimeInterval) {
-        super.update(currentTime)
-        
-        self.children.forEach { (child) in
-            if let updatable = child as? Updatable{
-                updatable.update(currentTime)
-            }
-        }
-    }
-    
-
-    
     @objc func movetoLobbyScene(_ sender:UIButton!)
     {
         
@@ -52,11 +41,4 @@ class WashScene: SKScene {
         NotificationCenter.default.post(name: Notification.Name.didRecieveButtonInput, object:nil)
         screenDelegate?.SwitchScreens(string: "LobbyScene")
     }
-
-    
-    
-    
-    
-    
-    
 }
